requirements.txt
# Runtime dependencies for the standalone CLI.
PyYAML==6.0.1

